refactor(tests): split dashscope provider errors
